tourpay/requirements/local.txt
django-debug-toolbar==3.1.1
django-extensions==3.0.9